deleted pfx file
