+++ NDB.Covid19/NDB.Covid19.Test/Tests/OAuth2/IdentityServerTest.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc.Testing;
 using Xunit;
@@ -21,93 +20,28 @@ namespace NDB.Covid19.Test.Tests.OAuth2
         {
             _factory = factory;
         }
-        /*public static X509Certificate2 CreateRsaCertificate(
-            string dnsName, int validityPeriodInYears)
-        {
-            var basicConstraints = new BasicConstraints
-            {
-                CertificateAuthority = false,
-                HasPathLengthConstraint = false,
-                PathLengthConstraint = 0,
-                Critical = false
-            };
-
-            var subjectAlternativeName = new SubjectAlternativeName
-            {
-                DnsName = new List<string> { dnsName }
-            };
-
-            var x509KeyUsageFlags = X509KeyUsageFlags.DigitalSignature;
-
-            var enhancedKeyUsages = new System.Security.Cryptography.OidCollection
-                {
-                    new Oid("1.3.6.1.5.5.7.3.1"),  // TLS Server auth
-                    new Oid("1.3.6.1.5.5.7.3.2"),  // TLS Client auth
-                };
-
-
-            var certificate = _cc.NewRsaSelfSignedCertificate(
-                new DistinguishedName { CommonName = dnsName },
-                basicConstraints,
-                new ValidityPeriod
-                {
-                    ValidFrom = DateTimeOffset.UtcNow,
-                    ValidTo = DateTimeOffset.UtcNow.AddYears(validityPeriodInYears)
-                },
-                subjectAlternativeName,
-                enhancedKeyUsages,
-                x509KeyUsageFlags,
-                new RsaConfiguration { KeySize = 2048 }
-            );
-
-            return certificate;
-        }
-        public void CreateStoreCert()
-        {
-            var sp = new ServiceCollection()
-                           .AddCertificateManager()
-                           .BuildServiceProvider();
-
-            _cc = sp.GetService<CreateCertificates>();
-
-            var rsaCert = CreateRsaCertificate("localhost", 10);
-
-            string password = "1234";
-            var iec = sp.GetService<ImportExportCertificate>();
-            
-            var rsaCertPfxBytes =
-                iec.ExportSelfSignedCertificatePfx(password, rsaCert);
-            //File.WriteAllBytes("rsaCert.pfx", rsaCertPfxBytes);
-            certificate = new X509Certificate2(rsaCertPfxBytes);
-
-            //X509Store store = new X509Store("teststore", StoreLocation.CurrentUser);
-            //store.Open(OpenFlags.ReadWrite);
-            //store.Add(certificate1);
-            //store.Close();
-        }*/
+        
         [Theory]
         [InlineData("/.well-known/openid-configuration/jwks")]
         public async Task Get_EndpointsReturnSuccessAndCorrectContentType(string url)
         {
             // Arrange
             AuthenticationManager authenticationManager = new AuthenticationManager();
-
-            Chilkat.Pfx pfx = new Chilkat.Pfx();
-            // This will get the current WORKING directory (i.e. \bin\Debug)
-            string workingDirectory = Environment.CurrentDirectory;
-            // This will get the current PROJECT directory
-            string projectDirectory2 = Directory.GetParent(workingDirectory).Parent.Parent.FullName;
-            _path = projectDirectory2 + "\\Tests\\OAuth2\\rsaCert.pfx";
-            
-            pfx.LoadPfxFile(_path, "1234");
-            //string path = Directory.GetCurrentDirectory();
             authenticationManager.client = _factory.CreateClient();
-            
+            File.WriteAllBytes("rsaCert.pfx", Startup.RsaCertPfxBytes);
+            Chilkat.Pfx pfx = new Chilkat.Pfx();
+            if (File.Exists("rsaCert.pfx"))
+            {
+                pfx.LoadPfxFile("rsaCert.pfx", "12345");
+            }
+           
+            File.Delete("rsaCert.pfx");
 
             // Act
             //public key from local cert
             string alias = "my_ecc_key1";
             string password = "secret123";
+            
             Chilkat.JavaKeyStore jks = pfx.ToJavaKeyStore(alias, password);
             Chilkat.StringBuilder sbJwkSet = new Chilkat.StringBuilder();
             jks.ToJwkSet(password, sbJwkSet);
